LayoutTests/platform/mac-sierra/editing/mac/attributed-string/attributed-string-for-typing-with-color-filter-expected.txt

Some text here
Input:
<div id="editor" style="-apple-color-filter: apple-invert-lightness(); color: rgba(20, 20, 20, 0.4); background: #ccc;" contenteditable="">Some text here</div>

Output:
[ ]
    NSBackgroundColor: #5c5c5c (NSCustomColorSpace)
    NSColor: rgba(239, 239, 239, 0.4) (NSCustomColorSpace)
    NSFont: Times-Roman 16.00 pt.
